backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==3.6.1
    # via starlette
bcrypt==4.0.0
    # via -r requirements.in
beautifulsoup4==4.11.1
    # via -r requirements.in
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via uvicorn
fastapi==0.81.0
    # via -r requirements.in
geographiclib==1.52
    # via geopy
geopy==2.2.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
html5lib==1.1
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   requests
lxml==4.9.1
    # via -r requirements.in
pydantic==1.10.1
    # via fastapi
pymongo==4.2.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
requests==2.28.1
    # via -r requirements.in
six==1.16.0
    # via
    #   html5lib
    #   python-dateutil
sniffio==1.3.0
    # via anyio
soupsieve==2.3.2.post1
    # via beautifulsoup4
starlette==0.19.1
    # via fastapi
typing-extensions==4.3.0
    # via pydantic
urllib3==1.26.12
    # via requests
uvicorn==0.18.3
    # via -r requirements.in
webencodings==0.5.1
    # via html5lib
